+++ R/modelsummary.R
@@ -14,10 +14,12 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #'   \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
 #'   \item When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kableExtra::save_kable`, or `cat` functions.
 #' }
-#' @param fmt string which specifies how numeric values will be rounded. This
-#' string is passed to the `sprintf` function. '\%.3f' will keep 3 digits after
-#' the decimal point with trailing zero. '\%.5f' will keep 5 digits. '\%.3e' will
-#' use exponential notation. See `?sprintf` for more options.
+#' @param fmt an integer, string, or function which determines how to format numeric values:
+#' \itemize{
+#'   \item integer: the number of digits to keep after the period (`format(round(x, fmt), nsmall=fmt)`)
+#'   \item character: string is passed to the `sprintf` function. '\%.3f' will keep 3 digits after the decimal point with trailing zero. '\%.5f' will keep 5 digits. '\%.3e' will use exponential notation. See `?sprintf` for more options.
+#'   \item function: a function which returns a formatted character string
+#' }
 #' @param stars to indicate statistical significance
 #' \itemize{
 #'   \item FALSE (default): no significance stars.
@@ -122,7 +124,7 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' @export
 modelsummary <- function(models,
                          output = "default",
-                         fmt = '%.3f',
+                         fmt = 3,
                          statistic = 'std.error',
                          statistic_override = NULL,
                          statistic_vertical = TRUE,